[editor] Move test utils into editor/shared/tests.
