+++ app/src/editor/canvas/tests/variadic.test.js
@@ -1,6 +1,7 @@
+import { setup, loadModuleDefinition } from '$editor/shared/tests/utils'
+
 import * as State from '../state'
 import * as Services from '../services'
-import { loadModuleDefinition, setup } from './utils'
 
 describe('Variadic Port Handling', () => {
     let teardown
